lib/pbio/platform/prime_hub_spike-rt: Revert Port F assignment.
Revert Port F assignment to a TOPPERS serial interface driver and
enable it to control a PUP device again.
This addresses a request from many users.

TODO:
Introduce a parameter that enables users to choose whether to assign Port F
to a TOPPERS serial interface driver or use it for controlling a PUP device.